🎨	코드의 구조/형태 개선 : Restful 하도록 컨트롤러를 단수와 복수로 변경
